test(runtime-host): parallelize execution host integration coverage
